API change: domain -> mesh
The domain (in mathematical terms) may refer to something with complicated
geometry, e.g. a sphere. A mesh may only be an approximation of this complicated
geometry. Consequently, it is more appropriate to use the term 'mesh' as the
parent representation of all n-cells. This way, one can have multiple meshes
for the same mathematical domain (different resolution, etc.) while keeping a concise language.