Updated build system to build without pedantic flags and tests in the release package.
Also updated a missing copy of doc/CMakeLists.txt to the release folder.
The rationale for disabling pedantic flags is that this leads to errors on Visual Studio, where -Werror cannot be translated properly.
A build of all tests, on the other hand, creates too many targets, intimitating new users.
Advanced users can still turn the option on via CMake.