Consistently disabled Visual Studio warning 4503 (truncated name decoration) to tests and examples.
Visual Studio complains about truncated type names exceeding 4096 characters.
With template metaprogramming, however, one exceeds this limit easily.
For users this is mostly harmless, hence we disable the warning in all examples and tests.
We do *not* disable this globally in the ViennaGrid headers, as this might interfere with other user configurations.